workspace/pipeline-output-servers/qwen-max-latest/mcp_screenshot_automation/refined/requirements.txt
mcp[cli]
pillow
fastapi
uvicorn
httpx
pydantic
jsonschema
python-multipart
jinja2
marko
typer
click
colorama
starlette
websockets
asyncpg
psycopg2
redis
motor
pyyaml
cryptography
python-jose
passlib
argon2-cffi
email-validator
pyotp
qrcode
zxing
opencv-python
numpy
pandas
matplotlib
seaborn
scikit-learn
tensorflow
torch
transformers
diffusers
onnx
onnxruntime
tesseract
pytesseract
selenium
playwright
beautifulsoup4
lxml
requests
scrapy
celery
redis
flower
sqlalchemy
alembic
pytest
pytest-cov
pytest-xdist
pytest-asyncio
black
isort
flake8
mypy
types-Pillow
types-requests
types-python-dateutil
types-toml
types-pkg_resources
types-setuptools
types-six
types-urllib3
types-werkzeug
types-flask
types-click
types-paramiko
types-pillow
types-psycopg2
types-pymysql
types-redis
types-sqlalchemy
types-ujson
types-websockets
types-pandas
types-numpy
types-matplotlib
types-scipy
types-sklearn
types-tensorflow
types-torch
types-transformers
types-diffusers
types-onnx
types-tesseract
types-selenium
types-playwright
types-beautifulsoup4
types-lxml
types-requests-html
types-scrapy
types-celery
types-flask-sqlalchemy
types-flask-login
types-flask-wtf
types-flask-bcrypt
types-flask-caching
types-flask-debugtoolbar
types-flask-migrate
types-flask-script
types-flask-security-too
types-flask-session
types-flask-socketio
types-flask-oidc
types-flask-openid
types-flask-jwt
types-flask-jwt-extended
types-flask-restful
types-flask-smorest
types-flask-apispec
types-flask-sqlalchemy-session
types-flask-testing
types-flask-mail
types-flask-admin
types-flask-babel
types-flask-flatpages
types-flask-freeze
types-flask-login
types-flask-principal
types-flask-restless
types-flask-security-too
types-flask-sijax
types-flask-sockets
types-flask-staticdigest
types-flask-stoid
types-flask-talisman
types-flask-themes2
types-flask-uploads
types-flask-wtf
types-flask-zurb-foundation
types-flask-influxdb
types-flask-paginate
types-flask-peewee
types-flask-pure
types-flask-pytest
types-flask-rq2
types-flask-scheduler
types-flask-sqlacodegen
types-flask-user
types-flask-webpack
types-flask-wtf
types-flask-zeroconf
types-flask-zurb-foundation
types-flask-influxdb
types-flask-paginate
types-flask-peewee
types-flask-pure
types-flask-pytest
types-flask-rq2
types-flask-scheduler
types-flask-sqlacodegen
types-flask-user
types-flask-webpack
types-flask-zeroconf